Dockerization.
Mongo and app WAR are copied as separate tasks in gradle, and buildImage depends on both.
Added oauth-credentials example XML that you personalize and is then copied in as override to Dockerfile.
Dockerfile fixed to copy in new WAR file to drive mongo.
Added bits to README to indicate specifying the CONCIERGE_URL, since you have to.